# openapi.yaml
openapi: 3.1.0
info:
  title: Persistio API
  version: 0.1.19
  description: Self-hosted memory service for AI agents.
servers:
  - url: http://localhost:4827
paths:
...
  /admin/vaults:
    get:
      summary: List vaults
      security:
        - adminKey: []
      responses:
        '200':
          description: Vault list
    post:
      summary: Create a vault
      security:
        - adminKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  minLength: 1
                purpose:
                  type: string
                  minLength: 1
                  maxLength: 500
                plan:
                  type: string
                  enum: [free, unlimited]
      responses:
        '201':
          description: Vault created
  /admin/vaults/{id}:
    get:
      summary: Get vault detail without key material
      security:
        - adminKey: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Vault detail
        '404':
          description: Vault not found
    patch:
      summary: Update a vault
      security:
        - adminKey: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              minProperties: 1
              properties:
                name:
                  type: string
                  minLength: 1
                purpose:
                  type: string
                  nullable: true
                  minLength: 1
                  maxLength: 500
                plan:
                  type: string
                  enum: [free, unlimited]
      responses:
        '200':
          description: Vault updated
        '404':
          description: Vault not found
    delete:
      summary: Delete a vault
      security:
        - adminKey: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Vault deleted
        '404':
          description: Vault not found
  /admin/vaults/{id}/rotate-key:
    post:
      summary: Rotate a vault API key
      security:
        - adminKey: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Vault key rotated
        '404':
          description: Vault not found
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    adminKey:
      type: apiKey
      in: header
      name: X-Admin-Key